+++ src/main/java/dk/kb/netarchivesuite/solrwayback/properties/PropertiesLoaderWeb.java
@@ -26,7 +26,16 @@ public class PropertiesLoaderWeb {
     public static final String WEBAPP_BASEURL_PROPERTY="webapp.baseurl";  //TODO DELETE WHEN FRONTEND HAS CHANGED NAME
     public static final String WEBAPP_PREFIX_PROPERTY="webapp.prefix";
     
-    public static final String OPENWAYBACK_SERVER_PROPERTY="openwayback.baseurl";	
+    //Deprecated. Use  PLAYBACK_ALTERNATIVE_ENGINE_PROPERTY instead. Backwards compatible, but will be removed in future version.
+    private static final String OPENWAYBACK_SERVER_PROPERTY="openwayback.baseurl";	
+    
+    public static final String PLAYBACK_ALTERNATIVE_ENGINE_PROPERTY="playback.alternative.engine";
+    
+    //Backwards compatible. Use 'playback.primary.engine' and 'playback.alternative.engine'
+
+    // This will overwrite the default SolrWayback playback if value set in property file.
+    public static final String PLAYBACK_PRIMARY_ENGINE_PROPERTY="playback.primary.engine"; 
+    
     public static final String ALTERNATIVE_PLAYBACK_COLLECTION_MAPPING_PROPERTY="alternative.playback.collection.mapping";
     
     public static final String FACETS_PROPERTY = "facets";	
@@ -63,15 +72,17 @@ public class PropertiesLoaderWeb {
     
     public static LinkedHashMap<String,String> ALTERNATIVE_PLAYBACK_COLLECTION_MAPPING= new LinkedHashMap<String,String>(); 
     public static String SOLRWAYBACK_VERSION; //Will be set from initialcontext-listener
-    public static String OPENWAYBACK_SERVER;
+    public static String OPENWAYBACK_SERVER; //Deprecated, to be removed in future version
+    public static String PLAYBACK_PRIMARY_ENGINE;
+    public static String PLAYBACK_ALTERNATIVE_ENGINE;
     public static int ARCHIVE_START_YEAR;
     public static String WAYBACK_SERVER = null;
     
     public static String WEBAPP_PREFIX = null;
     public static String MAPS_LATITUDE;
     public static String MAPS_LONGITUDE;
-    public static String MAPS_RADIUS;
-
+    public static String MAPS_RADIUS;   
+    
     public static int WARC_ENTRY_TEXT_MAX_CHARACTERS = 100*1024*1024; // 100 MB
 
     public static long EXPORT_CSV_MAXRESULTS=10000000;// 10M default
@@ -137,7 +148,17 @@ public class PropertiesLoaderWeb {
             STATS_NUMERIC_FIELDS = Arrays.asList(getProperty(NUMERIC_STATS_PROPERTY, StringUtils.join(STATS_NUMERIC_FIELDS, ",")).split(", *"));
             WORDCLOUD_STOPWORDS = Arrays.asList(getProperty(WORDCLOUD_STOPWORDS_PROPERTY, StringUtils.join(WORDCLOUD_STOPWORDS, ",")).split(", *"));
             WEBAPP_PREFIX = serviceProperties.getProperty(WEBAPP_PREFIX_PROPERTY,"/solrwayback/"); //Default to /solrwayback/ if not defined
-            OPENWAYBACK_SERVER = serviceProperties.getProperty(OPENWAYBACK_SERVER_PROPERTY);
+                                   
+            PLAYBACK_ALTERNATIVE_ENGINE = serviceProperties.getProperty(PLAYBACK_ALTERNATIVE_ENGINE_PROPERTY);
+
+            //BACKWARDS COMPATIBLE. Code will be removed in future version
+            if (PLAYBACK_ALTERNATIVE_ENGINE == null) {
+                PLAYBACK_ALTERNATIVE_ENGINE = serviceProperties.getProperty(OPENWAYBACK_SERVER_PROPERTY);                
+                if (PLAYBACK_ALTERNATIVE_ENGINE != null) {
+                  log.warn("Property:"+OPENWAYBACK_SERVER_PROPERTY +" is deprecated. Change to new property name:"+PLAYBACK_ALTERNATIVE_ENGINE_PROPERTY);                      
+                }            
+            }
+            
             MAPS_LATITUDE = serviceProperties.getProperty(MAPS_LATITUDE_PROPERTY);
             MAPS_LONGITUDE = serviceProperties.getProperty(MAPS_LONGITUDE_PROPERTY);
             MAPS_RADIUS = serviceProperties.getProperty(MAPS_RADIUS_PROPERTY);
@@ -163,6 +184,13 @@ public class PropertiesLoaderWeb {
             String warc_expanded_max_results= serviceProperties.getProperty(EXPORT_WARC_EXPANDED_MAXRESULTS_PROPERTY);
             String search_pagination= serviceProperties.getProperty(SEARCH_PAGINATION_PROPERTY);
 
+            PLAYBACK_PRIMARY_ENGINE = serviceProperties.getProperty(PLAYBACK_PRIMARY_ENGINE_PROPERTY);
+            if (PLAYBACK_PRIMARY_ENGINE == null) { //TODO delete after old variable has been deleted
+                
+                //PLAYBACK_PRIMARY_ENGINE=  serviceProperties.getProperty(PLAYBACK_PRIMARY_ENGINE);XXX
+            }
+            
+            
             if (csv_max_results != null) {
                 EXPORT_CSV_MAXRESULTS  = Long.parseLong(csv_max_results.trim());                
             }
@@ -224,9 +252,11 @@ public class PropertiesLoaderWeb {
             else {
              log.info("No collection playback mapping loaded.");   
             }
-            //Set max export sizes                                   
+           
+            
             log.info("Property:"+ WEBAPP_PREFIX_PROPERTY +" = " + WEBAPP_PREFIX);
-            log.info("Property:"+ OPENWAYBACK_SERVER_PROPERTY +" = " + OPENWAYBACK_SERVER);
+            log.info("Property:"+ PLAYBACK_ALTERNATIVE_ENGINE_PROPERTY +" = " + PLAYBACK_ALTERNATIVE_ENGINE);
+            log.info("Property:"+ PLAYBACK_PRIMARY_ENGINE_PROPERTY +" = " + PLAYBACK_PRIMARY_ENGINE);            
             log.info("Property:"+ ALLOW_EXPORT_WARC_PROPERTY +" = " + ALLOW_EXPORT_WARC);
             log.info("Property:"+ ALLOW_EXPORT_CSV_PROPERTY +" = " + ALLOW_EXPORT_CSV);
             log.info("Property:"+ WARC_ENTRY_TEXT_MAX_CHARACTERS_PROPERTY +" = " + WARC_ENTRY_TEXT_MAX_CHARACTERS);
@@ -249,7 +279,7 @@ public class PropertiesLoaderWeb {
             log.info("Property:"+ LEAFLET_SOURCE_PROPERTY +" = " + LEAFLET_SOURCE);
             log.info("Property:"+ LEAFLET_ATTRIBUTION_PROPERTY +" = " + LEAFLET_ATTRIBUTION);
             log.info("Property:"+ TOP_LEFT_LOGO_IMAGE_PROPERTY +" = " + TOP_LEFT_LOGO_IMAGE);
-            log.info("Property:"+ TOP_LEFT_LOGO_IMAGE_LINK_PROPERTY +" = " + TOP_LEFT_LOGO_IMAGE_LINK);
+            log.info("Property:"+ TOP_LEFT_LOGO_IMAGE_LINK_PROPERTY +" = " + TOP_LEFT_LOGO_IMAGE_LINK);            
             
             if (ALTERNATIVE_PLAYBACK_COLLECTION_MAPPING.size() >0) {
                 for (String key : ALTERNATIVE_PLAYBACK_COLLECTION_MAPPING.keySet())
@@ -261,7 +291,7 @@ public class PropertiesLoaderWeb {
         }
         catch (Exception e) {
             e.printStackTrace();
-            log.error("Could not load property file:"+ propertyFile);
+            log.error("Could not load property file:"+ propertyFile,e);            
             // TODO: This should be a catastrophic failure as the properties contains security oriented settings
         }